+++ entity/Bullet/SineSpeed.cs
@@ -3,18 +3,18 @@ using Godot;
 public partial class SineSpeed : BulletBasic 
 {
 	//Bullet that speed up and down by sine wave.
-	[Export] float frequency = 1;
-	protected float[] ages;
+	[Export] double frequency = 1;
+	protected double[] ages;
 
 	protected override void SortBullet()
 	{
 		ages[index] = ages[lastIndex];
 		base.SortBullet();
 	}
-	protected override void Move(in float delta)
+	protected override void Move(in double delta)
 	{
 		ages[index] += delta * frequency;
-		velocities[index] = velocities[index].Normalized() * speed * Mathf.Abs(Mathf.Sin(ages[index]));
+		velocities[index] = velocities[index].Normalized() * speed * (float) Mathf.Abs(Mathf.Sin(ages[index]));
 		
 		base.Move(delta);
 	}
